change modal cart

+++ backend/app/src/main/java/com/congthanh/project/repository/ecommerce/product/ProductCustomRepository.java
@@ -14,6 +14,6 @@ public interface ProductCustomRepository {
 
     boolean checkExistSlug(String slug);
 
-    int countTotalSoldProduct(String productId);
+    Long countTotalSoldProduct(String productId);
 
 }
